Adds `VersionManagerDelegate` to decouple the reloading of a new JS bundle from the developer's `App Delegate`
This allows us to remove `AppReloader` since it made assumptions about the developer's implementation (not everyone
has an app delegate named `AppDelegate`). This also allows the developer to define the way in which the new bundle
gets loaded (could be a new view controller or perhaps another view inside some other view controller - or maybe they
don't want a flip animation).

I opted to keep the `ViewController` there since it eliminates some boilerplate and is not coupled to the developer's
implementation. It's just a helper that can't be used if desired.